requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.1.4
python-docx==1.0.1
python-dotenv==1.0.0
html5lib==1.1
dash==2.14.2
dash-bootstrap-components==1.5.0
plotly==5.18.0
pandas==2.1.4
openpyxl==3.1.5
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
streamlit==1.35.0